+++ content/status.php
@@ -25,12 +25,17 @@
 			<?php wp_link_pages(); ?>
 		</div><!-- .entry-content -->
 
-		<footer class="entry-footer">
+		<div class="entry-byline small">
 			<?php hybrid_post_format_link(); ?>
+			<time <?php hybrid_attr( 'entry-published' ); ?>><?php echo get_the_date(); ?></time>
 			<span class="post-by"><?php echo esc_html_x( 'by', 'post author', 'themelia' ) ?></span>
 			<span <?php hybrid_attr( 'entry-author' ); ?>><?php the_author_posts_link(); ?></span>
-			<time <?php hybrid_attr( 'entry-published' ); ?>><?php echo get_the_date(); ?></time>
-			<?php edit_post_link(); ?><br />
+			<?php if ( comments_open() && 0 < $number  ) : '<span class="sep">' . _ex( ' | ', 'By Line separator', 'themelia' ) . '</span>'; endif ?>
+			<?php comments_popup_link( ( '' ), '<span>' . number_format_i18n( 1 ) . '</span>' . __( ' Comment', 'themelia' ), '<span>%</span>' .  __( ' Comments', 'themelia' ), 'comments-link' . $css_class, '' ); ?>
+			<?php edit_post_link(); ?>
+		</div><!-- .entry-byline -->
+
+		<footer class="entry-footer">
 			<?php hybrid_post_terms( array( 'taxonomy' => 'category', 'sep' => '<span>,</span> ',  'text' => esc_html__( 'Posted in: %s', 'themelia' ) ) ); ?>
 			<?php hybrid_post_terms( array( 'taxonomy' => 'post_tag', 'sep' => '<span>,</span> ', 'text' => esc_html__( 'Tagged: %s', 'themelia' ), 'before' => '<br />' ) ); ?>
 		</footer><!-- .entry-footer -->
@@ -49,7 +54,7 @@
 			<?php the_content(); ?>
 		</div><!-- .entry-content -->
 
-		<?php if ( ! get_option( 'show_avatars' ) ) : // If avatars are not enabled. ?>
+		<?php //if ( ! get_option( 'show_avatars' ) ) : // If avatars are not enabled. ?>
 
 			<footer class="entry-footer">
 				<?php hybrid_post_format_link(); ?>
@@ -60,7 +65,7 @@
 				<?php edit_post_link(); ?>
 			</footer><!-- .entry-footer -->
 
-		<?php endif; // End avatars check. ?>
+		<?php //endif; // End avatars check. ?>
 
 	<?php endif; // End single post check. ?>