+++ footer.php
@@ -1,10 +1,3 @@
-					<?php hybrid_get_sidebar( themelia_primary_sidebar('primary') ); // Calls themelia_primary_sidebar() function and loads the sidebar/*.php template. ?>
-				</div><!-- .inner .main-inner -->
-			</div><!-- .grid-container -->
-		</div><!-- #main -->
-
-		<?php hybrid_get_sidebar( 'subsidiary' ); // Loads the sidebar/subsidiary.php template. ?>
-
 		<?php do_action('themelia_before_footer_widgets'); ?>
 
 		<?php get_template_part( 'content/footer-widgets' ); // Loads the content/footer-widgets.php template. ?>