+++ CHANGELOG.md
@@ -3,6 +3,19 @@ Change Log
 
 All notable changes to Themelia will be documented in this file. This changelog adheres to [Semantic Versioning](http://semver.org/). Version format will look like `3.2.1` where `3` is the major release, `2` is the minor release, and `1` is the patch release.
 
+v1.0.10 - 2016-12-15
+--------------------
+
+### Changed
+
+-	Opening and closing divs for main container are moved from header.php and footer.php into index.php. Makes it easier creating templates.
+-	New screenshot.png
+
+### Updated
+
+-	CSS optimization
+-	Small change of post-byline in content templates
+
 v1.0.9 - 2016-12-13
 -------------------